third_party/blink/web_tests/fast/forms/input-common/input-intrinsic-size-expected.txt

This is a testharness.js-based test.
[FAIL] color: percent height
  assert_equals: expected 27 but got 14
[FAIL] color: min-height
  assert_equals: expected 27 but got 14
Harness: the test ran to completion.
